framework: - Rename frameworkVr to frameworkOvr. - Make FrameworkOvr timeStep a double. - Make Framework time a double (since it keeps incrementing for a long while).

+++ framework/framework.h
@@ -537,8 +537,8 @@ public:
 	void drawVirtualDesktop();
 
 	bool quitRequested;
-	float time;
-	float timeStep;
+	double time;
+	float timeStep; 
 	
 	bool waitForEvents;
 	bool fullscreen;
